chore: create microservice analytics; update new folder structure

+++ pkg/utils/config.go
@@ -2,11 +2,11 @@ package utils
 
 import "strconv"
 
-func ParsePort(portStr string) uint16 {
+func ParsePort(portStr string) int {
 	port, err := strconv.ParseUint(portStr, 10, 16)
 	if err != nil {
 
 		return 0
 	}
-	return uint16(port)
+	return int(port)
 }
